+++ CHANGELOG.md
@@ -25,6 +25,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-3671 Zen Mode Tooltip improvement (target glycemia level)
 - YLP-3687 Update insulin stats order
 - YLP-3690 Update basal tooltip style in Daily view
+- YLP-3700 Update time change tooltip style in Daily view
 - YLP-3702 Update Time In Loop Mode stat style
 - YLP-3710 Move the "EVENTS" section above the glycemia curve
 - YLP-3730 Delete "Avg. Daily" in dashboard and trends view to lighten all the frame